#11 - move meta entities outside domain package - add json 2 pojo maven plugin to generate classes from metadata schema - cleanup entities from builder/constructor annotations - extend ReferenceAttributeEntity from base AttributeEntity
