+++ data-sync-repository/src/main/java/com/github/zybercik00/repository/mappingattribute/AttributeRepo.java
@@ -1,6 +1,6 @@
 package com.github.zybercik00.repository.mappingattribute;
 
-import com.github.zybercik00.domain.mappingAttribute.AttributeEntity;
+import com.github.zybercik00.domain.metadata.AttributeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AttributeRepo extends JpaRepository<AttributeEntity, String> {
