+++ data-sync-repository/src/main/java/com/github/zybercik00/repository/mappingattribute/MappingRepo.java
@@ -1,6 +1,6 @@
 package com.github.zybercik00.repository.mappingattribute;
 
-import com.github.zybercik00.domain.mappingAttribute.MappingEntity;
+import com.github.zybercik00.metadata.MappingEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MappingRepo extends JpaRepository<MappingEntity, String> {
